requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
PyPDF2==3.0.1
pandas==2.2.2
openpyxl==3.1.4
pydantic==2.7.1